Show class for LV2 effects when selecting
This provides a convenient sorting and better discoverability for LV2
effects, especially if more than a few are enabled from the web UI.

Configured plugins are grouped by categories and then sorted. The
categories are displayed as non-selectable headers in a different
color.

The built-in synths are still displayed first in their original order.