Show class for LV2 effects when selecting
This provides a convenient sorting and better discoverability for LV2
effects, especially if more than a few are enabled from the web UI.

For the moment, the class is displayed only as a prefix, but it might be
better to create nice sub-headings for these, as is done on other
screens.

It'd also be nice to eventually categorize the webconf this way as well.