Don't sleep when restoring ZS3 presets for LV2
In LV2, the plugin state is (usually?) fully set by setting the
controller values. So there's no need to set a preset first, nor to
sleep between setting the preset and then setting controller values.

By not doing this, ZS3 switching for effect chains with only LV2 are now
instant.