Configure input capture channels for a layer
This commit adds the ability to configure which capture inputs route to
the top input on an effects layer.

Internally, this is stored in a "audio_in" dict that precisely specifies
which capture inputs are connected to which LV2 input(s).

To the user, a simplified list of mono / stereo choices are provided,
depending on the number of capture inputs and number of LV2 inputs.