Fixes touchscreen pattern editor issues.
Touching anywhere in an existing note now selects the note.
Can adjust length of notes properly, not just from first step of note.
Click on last cell of a note (which is longer than one step) in would toggle next cell, not clicked cell.
Dificult to remove note on last step in pattern with touch. Need to touch left edge of last box.