modui: override __del__ to do nothing
__del__ is called during garbage collection, which may be some time
after an engine is stopped, and the default implementation of __del__
also calls self.stop().

This causes a problem in the following scenario:
- Running a mod-ui already
- Navigate to Main -> Clean All
- Run mod-ui again from a snapshot
- At some point (when garbage collection is triggered), mod-ui is stopped

I really think __del__ should not stop by default, due to the
unpredictability of when garbage collection gets triggered on an
already stopped engine. But this change just fixes that for the mod-ui
engine.