Fix logic for enable_touch_controller_switches
When there are visible screen buttons, the controller touch interface doesn't
need to be overloaded with button press actions. Accidental button presses are
common when making small touch value changes, so if there is buttons on the
actual screen the controllers shouldn't handle touches as buttons.

This seems to be the intent of enable_touch_controller_switches but the logic
is inverted and also includes a Z2 wiring layout which may no longer be
relevant.

This change also enables touching the top bar for back button presses,
regardless of the value of enable_touch_controller_switches. This is because
there is often no on screen alternative to trigger back button presses.