Fix race condition loading plugins with both controllers and presets.
The race condition can happen when restoring a snapshot, and a plugin
both has a preset and controller values. What is supposed to happen is
this:

1. Jalv is told to load a preset.

2. Jalv prints the controllers associated with that preset, which is
   picked up on the process pipe. This causes an update in the UI.

3. New controller values are set based on what was stored in the
   snapshot, which updates the plugin, and again the UI.

Because the the parsing of the Jalv messages in point 2 happens in a
separate thread, it's possible for step 2 and 3 to be reversed. In
this case it will cause the UI to go out of sync with the plugin,
since the printout from Jalv contains the values from the preset, not
the ones that were set most recently.

At that moment it is only a display issue, since the plugin will
actually be using the correct values. But the values in the UI are the
ones that are saved, so if the snapshot is saved again, the plugin
parameters will be wrong the next time the snapshot is loaded.

Fix this by forcing another controller printout from Jalv if we update
any controller values while we are still waiting for one.

Fixes:
https://github.com/zynthian/zynthian-issue-tracking/issues/1453

Signed-off-by: Kristian Amlie <kristian@amlie.name>